Move OasAgent::Agent to correct place
