+++ lib/oas_agent/control.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 # frozen_string_literal: true
 
-require "agent/agent"
+require "oas_agent/agent"
 require "agent/logger"
 require "oas_agent/agent_context"
